# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Api Virtual Unimed",
    "description": "Essa Api tem como objetivo mostra rotas e parametros usandos no app",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/"
    }
  ],
  "components": {
    "schemas": {
      "change_pass": {
        "type": "boolean",
        "properties": {
          "password": {
            "type": "string"
          },
          "comparepass":{
            "type": "string"
          }
        }
      },
      "services":{
        "type": "object",
        "properties": {
          "title":{
            "type": "string"
          },
          "description":{
            "type": "string"
          }
        }
      },
      "banners":{
        "type": "object",
        "properties": {
          "title":{
            "type": "string"
          },
          "image":{
            "type": "string"
          }
        }
      },
      "generation_code":{
        "type": "object",
        "properties": {
          "code":{
            "type": "string"
          }
        }
      },
      "code_validation":{
        "type": "object",
        "properties": {
          "code":{
            "type": "string"
          }
        }
      },
      "validation_user": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "phone":{
            "type":"string"
          }
        }
      }
    }
  },
  "paths": {
    "/change_pass": {
      "put": {
        "summary": "Fazer alteração da senha",
        "description": "Essa rota será responsável por fazer a alteração da senha",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/change_pass"
              },
              "examples": {
                "password": {
                  "value": {
                    "password": "1234ab",
                    "compare_pass": "1234ab"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Senha invalida"
          },
          "200": {
            "description": "Sucesso",
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "type": "boolean",
                    "$ref": "#/components/schemas/change_pass"
                }
              }
            }
          }
        }
      }
    },
    "/services": {
      "get": {
        "summary": "Fazer listagem dos serviços",
        "description": "Essa rota será responsável por listar os serviços da aplicação",
        "parameters": [
          {
            "name": "services",
            "in": "params",
            "description": "parâmetros da rota para listagem",
            "require": true
          }
        ],
        "responses": {
          "400": {
            "description": "Error"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "type": "object",
                    "$ref": "#/components/schemas/services"
                }
              }
            }
          }
        }
      }
    },
    "/banner": {
      "get": {
        "summary": "Fazer listagem dos banner",
        "description": "Essa rota será responsável por listar os banners da aplicação",
        "parameters": [
          {
            "name": "banner",
            "in": "params",
            "description": "parâmetros da rota para listagem",
            "require": true
          }
        ],
        "responses": {
          "400": {
            "description": "Error"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "type": "object",
                    "$ref": "#/components/schemas/banners"
                }
              }
            }
          }
        }
      }
    },
    "/generation_code": {
      "get": {
        "summary": "Fazer a criação do código de verificação",
        "description": "Essa rota será responsável por criar do código de verificação",
        "parameters": [
          {
            "name": "generation_code",
            "in": "params",
            "description": "parâmetros da rota de criação",
            "require": true
          }
        ],
        "responses": {
          "400": {
            "description": "Error"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "type": "object",
                    "$ref": "#/components/schemas/banners"
                }
              }
            }
          }
        }
      }
    },
    "/code_validation": {
      "post": {
        "summary": "Fazer a validação do código e alterar a senha",
        "description": "Essa rota será responsável por fazer a validação do código e alterar a senha",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/code_validation"
              },
              "examples": {
                "password": {
                  "value": {
                    "password": "1234ab",
                    "comparepass": "1234ab"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Código invalido"
          },
          "200": {
            "description": "Sucesso",
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "type": "boolean",
                    "$ref": "#/components/schemas/code_validation"
                }
              }
            }
          }
        }
      }
    },
    "/validation_user": {
      "get": {
        "summary": "Fazer validação do usuário",
        "description": "Essa rota será responsável por fazer a validação com cpf se usuário existe e retornar dados essenciais para prosseguir com fluxo",
        "parameters": [
          {
            "name": "cpf",
            "in": "query",
            "description": "parâmetros da rota de validação do usuário",
            "require": true
          }
        ],
        "responses": {
          "400": {
            "description": "cpf invalido"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "type": "object",
                    "$ref": "#/components/schemas/validation_user"
                }
              }
            }
          }
        }
      }
    }
  }
}